이수정 20240327 "Q&A 답변,목록 및 리뷰"

+++ src/main/java/com/suje/domain/customer/QnAVO.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class QnAVO {
-	long qna_code;
+	int qna_code;
 	String m_id, qna_title, qna_content, qna_reply, qna_qdate, qna_adate;
 }
